+++ src/pages/Main/Promo/Promo.css
