+++ src/pages/Main/Promo/Promo.js
