+++ src/pages/Main/Techs/Techs.css
