+++ src/pages/Main/Techs/Techs.js
