+++ src/pages/Movies/MovieCard/MovieCard.css
