+++ src/pages/index.js
@@ -1,0 +1,7 @@
+export { Main } from './Main/Main';
+export { Movies } from './Movies/Movies';
+export { SavedMovies } from './SavedMovies/SavedMovies';
+export { Profile } from './Profile/Profile';
+export { Register } from './Auth/Register/Register';
+export { Login } from './Auth/Login/Login';
+export { PageNotFound } from './PageNotFound/PageNotFound';
